Make PieceName class as a file to make it public seperate from Piece.java file
