+++ src/main/java/chess/pieces/Piece.java
@@ -75,23 +75,4 @@ public class Piece {
     }
 }
 
-enum PieceName {
-    PAWN('p'),
-    KNIGHT('n'),
-    ROOK('r'),
-    BISHOP('b'),
-    QUEEN('q'),
-    KING('k');
-
-    private final char tempRepresentation;
-
-    PieceName(char representation) {
-        this.tempRepresentation = representation;
-    }
-
-    public char getTempRepresentation() {
-        return tempRepresentation;
-    }
-}
-
